WIP: feat: Adds a notion of recipient to documents
This commit adds a notion of (optional) recipient to documents.

If a document has a recipient set it will only be displayed to this
recipient.

Current implementation uses a an HTTP header to tell the application
what is the (recipient) name of the user currently using the
application.

The goal of this feature is to create a secret addressed to a single
person, send the share URL via un-reliable support (e.g. e-mail) and
rely on a third-party SSO system to filter access to the secret.